Polishing, S01E01
- Remove dead code `LenderPool`
- Remove dead code `GuiTask`
- Change `TorrentFile`, `Subpiece` and `Piece` int types to `int64_t`
- Rework `encode_integer` and `decode_integer` to use `int64_t`
- Rework `socket` to use `int64_t` in `read` function
- Remove dead code `display_decode_error`
- Rework `Message` to use `int64_t`
- Rework `Bitfield` to use `int64_t`

+++ src/download/socket.hpp
@@ -41,7 +41,7 @@ class Socket {
   /// timeout.
   Outcome<SocketError> write(const std::vector<uint8_t>& buf, timeout timeout);
   /// Attempt to readn `n` bytes from the socket with the given timeout.
-  Result<std::vector<uint8_t>, SocketError> read(uint32_t n, timeout timeout);
+  Result<std::vector<uint8_t>, SocketError> read(int64_t n, timeout timeout);
 
   /// Close the socket
   Outcome<SocketError> close();
